Add ERC721ALowCap extension/remove existing tokensOfOwner created previously
